# Add sync.Pool to gzip middleware