Using a `sync.Pool` and the `gzip.Writer.Reset` method, we are able to reduce allocations, memory usage, cpu time and gc time. Benchmark below:

```
benchmark         old ns/op     new ns/op     delta
BenchmarkGzip     321829        94806         -70.54%

benchmark         old allocs     new allocs     delta
BenchmarkGzip     79             58             -26.58%

benchmark         old bytes     new bytes     delta
BenchmarkGzip     1459072       8065          -99.45%
```